requirements.txt
fastapi
uvicorn
requests
pydantic
python-dotenv
pytest
pytest-cov
httpx
